pyproject.toml
[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "core*"]
namespaces = false

[project]
name = "vox-hub"
version = "0.1.0"
scripts = {api = "api.main:main"}

dependencies = [
    "google-genai==1.29.0",
    "google-adk==1.8.0",
    "aiofiles==24.1.0",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "python-dotenv==1.1.1",
    "websockets==15.0.1",
    "pydantic-settings==2.5.2"
]
